+++ src/main/java/com/ilta/solepli/domain/place/repository/SolmarkPlaceRepository.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.ilta.solepli.domain.place.entity.SolmarkPlace;
+import com.ilta.solepli.domain.solmark.place.entity.SolmarkPlace;
 import com.ilta.solepli.domain.user.entity.User;
 
 @Repository
