+++ src/main/java/com/ilta/solepli/domain/solmap/service/SolmapService.java
@@ -24,7 +24,6 @@ import com.ilta.solepli.domain.category.repository.CategoryRepository;
 import com.ilta.solepli.domain.place.entity.Place;
 import com.ilta.solepli.domain.place.entity.QPlace;
 import com.ilta.solepli.domain.place.entity.QPlaceHour;
-import com.ilta.solepli.domain.place.entity.SolmarkPlace;
 import com.ilta.solepli.domain.place.entity.mapping.PlaceCategory;
 import com.ilta.solepli.domain.place.entity.mapping.QPlaceCategory;
 import com.ilta.solepli.domain.place.repository.PlaceRepository;
@@ -36,6 +35,7 @@ import com.ilta.solepli.domain.review.repository.ReviewRepository;
 import com.ilta.solepli.domain.review.repository.ReviewTagCustomRepository;
 import com.ilta.solepli.domain.solmap.dto.*;
 import com.ilta.solepli.domain.solmap.entity.SearchType;
+import com.ilta.solepli.domain.solmark.place.entity.SolmarkPlace;
 import com.ilta.solepli.domain.tag.entity.TagType;
 import com.ilta.solepli.domain.user.entity.User;
 import com.ilta.solepli.domain.user.util.CustomUserDetails;
